+++ hollow/src/test/java/com/netflix/hollow/api/producer/HollowMultiPublisherTest.java
@@ -22,6 +22,9 @@ import com.netflix.hollow.api.consumer.InMemoryBlobStore;
 import com.netflix.hollow.api.objects.generic.GenericHollowObject;
 import com.netflix.hollow.api.producer.HollowProducer.Populator;
 import com.netflix.hollow.api.producer.HollowProducer.WriteState;
+import com.netflix.hollow.api.producer.experimental.AbstractHollowMultiPublisher;
+import com.netflix.hollow.api.producer.experimental.ParallelHollowMultiPublisher;
+import com.netflix.hollow.api.producer.experimental.SingleHollowMultiPublisher;
 import com.netflix.hollow.api.producer.fs.HollowInMemoryBlobStager;
 import com.netflix.hollow.core.index.HollowPrimaryKeyIndex;
 import com.netflix.hollow.core.write.objectmapper.HollowPrimaryKey;
